Ooooh, getting closer to running user-defined words - see notes below
* `2DROP` seems to work fine; `SQUARE` is buggy, and `ROT` and `-ROT` are flat-out busted.
I think what's going on is the way stuff is getting pushed onto the return stack.  I'm gonna
try and walk thru it here.  So in other Forths, when I type:

	1 2 3 ROT .S

I get:

	2 3 1

But in Pa'lante, I get:

	2 0 2

So let's try and depuzzle this:

Step 1: User input is pushed onto the return stack.  So the return stack should be:

	1
	2
	3
	rot
	.s

So now when it runs:

	Numbers - they get pushed onto the stack; stack is 1 2 3
	rot - Ahhhh, here's the bug.  "2 roll" gets added AT THE END of the return stack.
	So to follow that out:
	.s - should print 1 2 3 (where's it getting 2 0 2?)
	2 roll...

Okay, so no.  That's not it.  But I still have an idea I want to try.  Basically a "shift onto bottom of stack";
like a "find-&-replace" - i.e. when it sees:
	rot
	.s
and it hits ROT, it changes the return stack to:
	2
	roll
	.s
Basically replacing "rot" with "2 roll", which was what I was shooting for