I did a huge rewrite! :)
* Sped up the interpreter by not decoding every word every step of the way (the main reason for the rewrite)
* Stopped overusing the return stack, and with that change in place, added BEGIN and AGAIN - my first branching words! :)
* And found I can define a lot of words in Forth - see backup.txt

This is a squash commit of a branch not on my GitHub, but the README and backup.txt should give you the gist of what's going on here.  Next on the to-do list:

* Re-add = (didn't have time to add it)
* Check if the following can be written in Forth, and re-add them if not: / MOD NEGATE INVERT
* Add comments and UNTIL
* Hold off on the following until I get the compiler working: ." S" COUNT TYPE

Once these are done, release version 0.3 - just the optimization BEGIN/AGAIN/UNTIL is more than enough for another zero-point release. :)